Run some sanity tests in %check
The `cctest` binary will crash when built with GCC 11 (with LTO enabled), and
it is the subject of the upstream bug report linked in the previous commit, so
make sure `cctest --help` returns success as a quick sanity test.

For Clear Linux OS, a test detected that `node --help` crashes after being
built in the same way, so test that command as well.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>